build: adopt the taskfile CI contract and fleet lint baseline (#55)
* build: adopt the taskfile CI contract with local includes

* build: adopt fleet lint configs and a managed gitignore block

* build: bump the Go directive to 1.25.12 for GO-2026-5856

* refactor: apply golangci-lint modernize fixes

* refactor: satisfy the remaining golangci-lint baseline findings

* style: tidy comments left by the lint fixes

+++ internal/oidc/device.go
@@ -76,7 +76,7 @@ func (m *Manager) DeviceStart(ctx context.Context) (DeviceAuth, error) {
 		VerificationURI:         resp.VerificationURI,
 		VerificationURIComplete: resp.VerificationURIComplete,
 		ExpiresAt:               expiry.Unix(),
-		Interval:                int64(resp.Interval),
+		Interval:                resp.Interval,
 	}, nil
 }
